Finished cutting up the subroutines from ice-Ih-gen.f.
Cut up all of the different subroutines in ice-Ih-gen.f and put them
into different *.f files in subroutines/. All of them compile and run
with f2py successfully. Still need to check and see if functionality is
available in python, but it probably is.

Next step: Assemble every piece in python code so that we can eventually
have the exact same output as the vanilla script does.