Managed to make one of the subroutines take arguments.
By removing the seemingly unnecessary beginning of the program where it
calls all of the various subroutines, I was able to make cons() take a
couple of arguments. It's a hack right now -- it just takes the
arguments and sets the COMMON block analogue to it. Works, though!

Next step: performing the same duty with the rest of the subroutines.